requirements.txt
flask
pandas
requests
python-telegram-bot
